LayoutTests/imported/w3c/web-platform-tests/workers/modules/dedicated-worker-options-credentials-expected.txt

Blocked access to external URL http://www1.localhost:8800/workers/modules/resources/export-credentials.py
CONSOLE MESSAGE: Importing a module script failed.

Harness Error (FAIL), message = Script error.

PASS Test initialization: setting up cross-origin cookie
PASS new Worker() with type=module and default credentials option should behave as credentials=same-origin and send the credentials
PASS new Worker() with type=module and credentials=omit should not send the credentials
PASS new Worker() with type=module and credentials=same-origin should send the credentials
PASS new Worker() with type=module and credentials=include should send the credentials
PASS new Worker() with type=module and default credentials option should behave as credentials=same-origin and send the credentials for same-origin static imports
PASS new Worker() with type=module and credentials=omit should not send the credentials for same-origin static imports
PASS new Worker() with type=module and credentials=same-origin should send the credentials for same-origin static imports
PASS new Worker() with type=module and credentials=include should send the credentials for same-origin static imports
TIMEOUT new Worker() with type=module and default credentials option should behave as credentials=same-origin and not send the credentials for cross-origin static imports Test timed out
NOTRUN new Worker() with type-module credentials=omit should not send the credentials for cross-origin static imports
NOTRUN new Worker() with type=module and credentials=same-origin should not send the credentials for cross-origin static imports
NOTRUN new Worker() with type=module and credentials=include should send the credentials for cross-origin static imports
NOTRUN new Worker() with type=module and default credentials option should behave as credentials=same-origin and send the credentials for same-origin dynamic imports
NOTRUN new Worker() with type=module and credentials=omit should not send the credentials for same-origin dynamic imports
NOTRUN new Worker() with type=module and credentials=same-origin should send the credentials for same-origin dynamic imports
NOTRUN new Worker() with type=module and credentials=include should send the credentials for same-origin dynamic imports
NOTRUN new Worker() with type=module and default credentials option should behave as credentials=same-origin and not send the credentials for cross-origin dynamic imports
NOTRUN new Worker() with type-module credentials=omit should not send the credentials for cross-origin dynamic imports
NOTRUN new Worker() with type=module and credentials=same-origin should not send the credentials for cross-origin dynamic imports
NOTRUN new Worker() with type=module and credentials=include should send the credentials for cross-origin dynamic imports
NOTRUN new Worker() with type=classic should always send the credentials regardless of the credentials option (default).
NOTRUN new Worker() with type=classic should always send the credentials regardless of the credentials option (omit).
NOTRUN new Worker() with type=classic should always send the credentials regardless of the credentials option (same-origin).
NOTRUN new Worker() with type=classic should always send the credentials regardless of the credentials option (include).
NOTRUN new Worker() with type=classic should always send the credentials for same-origin dynamic imports regardless of the credentials option (default).
NOTRUN new Worker() with type=classic should always send the credentials for same-origin dynamic imports regardless of the credentials option (omit).
NOTRUN new Worker() with type=classic should always send the credentials for same-origin dynamic imports regardless of the credentials option (same-origin).
NOTRUN new Worker() with type=classic should always send the credentials for same-origin dynamic imports regardless of the credentials option (include).
NOTRUN new Worker() with type=classic should never send the credentials for cross-origin dynamic imports regardless of the credentials option (default).
NOTRUN new Worker() with type=classic should never send the credentials for cross-origin dynamic imports regardless of the credentials option (omit).
NOTRUN new Worker() with type=classic should never send the credentials for cross-origin dynamic imports regardless of the credentials option (same-origin).
NOTRUN new Worker() with type=classic should never send the credentials for cross-origin dynamic imports regardless of the credentials option (include).
